User is now automatically appended to commit message, with special syntax.
No more expansion of %user in commit messages in ascription configuration,
nor separate commit messages with and without user.
User is added to both commit messages of normal and ascription catalogs.

In case the commit message is not given in config or in command line,
(or is empty), an editor is launched to input it.

svn path=/trunk/l10n-support/pology/; revision=994162