Deterministic order of catalogs in inverse mapping
The order of messages in summit catalog depends on the order
of catalogs in inverse mapping in the case of messages that
cannot be inserted by equality or source reference equality.
These messages are just appended to end, and therefore their
order depends on the order of gathered branch catalogs.

svn path=/trunk/l10n-support/pology/; revision=1413734