Write file with incomplete catalogs unconditionally
If output of list of incomplete catalogs to a file is requested,
this file should be written regardless if there were any incomplete
catalogs or not. If the file weren't written, it could happen that
an older version of the file from a prior run of the sieve is mistaken
for the result of the current run.

svn path=/trunk/l10n-support/pology/; revision=1446576