Regenerated exclusion dictionary files + updated script
The script has been updated to work with recent versions
of Norsk ordbank (and the ‘ordbanken’ software). It
now requires three versions of the word files in ‘ordbanken’,
revision 103, revision 183 and the most recent revision.

Revision 103 contains ‘klammeformer’ words (which are
removed in later versions). Revision 183 contains
‘unormert’ words (which are removed in later verions)
and some words which have been removed/fixed in later
versions.

The script now also generates separte ‘klammeformer’
and ‘unormert’ word files, and the rules have been
updated to reflect this. (A new ‘unormert.dat’ file
has been added for the ‘unormert’ words.)

svn path=/trunk/l10n-support/pology/; revision=1534485