Revert to hashing messages by their object id
On the one hand side in the Python 2 version it was assumed
that messages hash by their object id (i.e. memory pointer),
so two __eq__ messages may have had different hash values.
On the other hand, the meaning of __eq__ is very specific in
the context of a PO message and it is not equivalent to
simple equality of all attributes. (In retrospect it might
have been not the best design choice to at all allow use of
equality operator on PO messages.)
Therefore revert to using object id for hash.
This also recovers the intended semantics of __eq__.