pomtrans: remove attributes from HTML break elements used as Apertium engine separators
Issue #2 describes a problem where the (HTML-based) separator tags supplied by the `pomtrans` script are reformatted by the Apertium engine tools; the result is that the separator string isn't found in the translated response text.

In particular it was conversion of single-quotes (provided by `pomtrans`) to double-quotes (responded by Apertium) that caused the problem.

This changeset removes the use of an HTML attribute in the break elements in the separator, leaving only lowercase characters and less-than/greater-than symbols.

I've tested this locally using Python2, although at the time of writing there isn't any test coverage provided here.

Resolves #2.