CI, cmake test: include also non-default python dirs
A few files are installed outside the canonical python
locations, so they can't be found by tests executed
against the version of pology installed by cmake.
Force the value of CMAKE_INSTALLAT_PATH so that
the value of that directory can't be subject to
local config, and PYTHONPATH can be built from
that value.

The long term solution to handle the deployment is
to replace cmake with native python packaging.