[gl] Sort and format rules in lang/gl/rules/trasno.rules
This commit tries to unify the file structure, formats how each rule entry must be written,
formats the rule names in the lang/gl/rules/trasno.rules file.

The rule entry format must be as shown below:
- regexp
- id
- hint
- valid entries.

We sort all the rules by two criteria:
- Sort by Trasnada edition (the Trasno's event to coordinate translators
  with aggrements)
- Sort alphabetically by rule name

Please check https://termos.trasno.gal/ searching for Trasnada to see
the list of rules that should be in that file.